Rename instructions_new.rs to instructions.rs
