+++ src/lib.rs
@@ -23,7 +23,6 @@ pub mod method_handle;
 pub mod code_item;
 pub mod opcodes;
 pub mod instructions;
-pub mod instructions_new;
 pub mod disasm;
 
 /* Actually unused for now but there should
